requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic
beautifulsoup4==4.12.3
mcp[cli]==1.3.0
httpx==0.27.0 